+++ crates/symbolicator-proguard/src/symbolication.rs
@@ -15,6 +15,12 @@ impl ProguardService {
     ///
     /// "Symbolicate" here means that exceptions and stack
     /// frames are remapped using proguard files.
+    #[tracing::instrument(skip_all,
+        fields(
+            exceptions = request.exceptions.len(),
+            frames = request.stacktraces.iter().map(|st| st.frames.len()).sum::<usize>())
+        )
+    ]
     pub async fn symbolicate_jvm(
         &self,
         request: SymbolicateJvmStacktraces,
@@ -184,93 +190,164 @@ impl ProguardService {
         frame: &JvmFrame,
         release_package: Option<&str>,
     ) -> Vec<JvmFrame> {
-        let stack_frame = frame
+        let deobfuscated_signature: Option<proguard::DeobfuscatedSignature> =
+            frame.signature.as_ref().and_then(|signature| {
+                for mapper in mappers {
+                    if let Some(deobfuscated_signature) = mapper.deobfuscate_signature(signature) {
+                        return Some(deobfuscated_signature);
+                    }
+                }
+                None
+            });
+
+        let params = deobfuscated_signature
+            .as_ref()
+            .map(|sig| sig.parameters_types().collect::<Vec<_>>().join(","));
+
+        // We create the proguard frame according to these priorities:
+        // * Use the frame's line number if it exists
+        // * Use the frame's parameters if they exist
+        // * Use line number 0
+        let proguard_frame = frame
             .lineno
             .map(|lineno| {
                 proguard::StackFrame::new(&frame.module, &frame.function, lineno as usize)
             })
             .or_else(|| {
-                frame.parameters.as_ref().map(|params| {
+                params.as_ref().map(|p| {
                     proguard::StackFrame::with_parameters(
                         &frame.module,
                         &frame.function,
-                        params.as_str(),
+                        p.as_str(),
                     )
                 })
-            });
+            })
+            // This is for parity with the Python implementation. It's unclear why remapping a frame with line 0
+            // would produce useful information, and I have no conclusive evidence that it does.
+            // See the `line_0_1` and `line_0_2` unit tests in this file for examples of the results this produces.
+            //
+            // TODO(@loewenheim): Find out if this is useful and remove it otherwise.
+            // The PR that introduced this was https://github.com/getsentry/symbolicator/pull/1434.
+            //
+            // UPDATE(@loewenheim): The retrace implementation at https://dl.google.com/android/repository/commandlinetools-mac-11076708_latest.zip
+            // returns the same value whether you give it line 0 or no line at all, and it is the same result that our implementation
+            // gives with line 0. This indicates that the _behavior_ is correct, but we should be able to get there without
+            // backfilling the line number with 0.
+            .unwrap_or_else(|| proguard::StackFrame::new(&frame.module, &frame.function, 0));
 
         // First, try to remap the whole frame.
-        // This only works if it has a line number or params.
-        if let Some(proguard_frame) = stack_frame {
-            let mut mapped_frames = Vec::new();
-
-            for mapper in mappers {
-                mapped_frames.clear();
-                mapped_frames.extend(mapper.remap_frame(&proguard_frame));
-
-                if mapped_frames.is_empty() {
-                    continue;
-                }
-
-                let bottom_class = mapped_frames[mapped_frames.len() - 1].class();
-
-                // sentry expects stack traces in reverse order
-                return mapped_frames
+        let mut mapped_frames = Vec::new();
+        let mut frames = mappers
+            .iter()
+            .find_map(|mapper| {
+                Self::map_full_frame(mapper, frame, &proguard_frame, &mut mapped_frames)
+            })
+            // Second, try to remap the frame's method.
+            .or_else(|| {
+                mappers
                     .iter()
-                    .rev()
-                    .map(|new_frame| {
-                        let mut mapped_frame = JvmFrame {
-                            module: new_frame.class().to_owned(),
-                            function: new_frame.method().to_owned(),
-                            lineno: Some(new_frame.line() as u32),
-                            ..frame.clone()
-                        };
+                    .find_map(|mapper| Self::map_class_method(mapper, frame))
+            })
+            // Third, try to remap just the frame's class.
+            .or_else(|| {
+                mappers
+                    .iter()
+                    .find_map(|mapper| Self::map_class(mapper, frame))
+            });
 
-                        // clear the filename for all *foreign* classes
-                        if mapped_frame.module != bottom_class {
-                            mapped_frame.filename = None;
-                            mapped_frame.abs_path = None;
-                        }
-
-                        // mark the frame as in_app after deobfuscation based on the release package name
-                        // only if it's not present
-                        if let Some(package) = release_package {
-                            if mapped_frame.module.starts_with(package)
-                                && mapped_frame.in_app.is_none()
-                            {
-                                mapped_frame.in_app = Some(true);
-                            }
-                        }
-                        mapped_frame
-                    })
-                    .collect();
-            }
-        }
-
-        // Second, if that is not possible, try to re-map only the class-name.
-        for mapper in mappers {
-            let Some(mapped_class) = mapper.remap_class(&frame.module) else {
-                continue;
-            };
-
-            let mut mapped_frame = JvmFrame {
-                module: mapped_class.to_owned(),
-                ..frame.clone()
-            };
-
-            // mark the frame as in_app after deobfuscation based on the release package name
-            // only if it's not present
-            if let Some(package) = release_package {
-                if mapped_frame.module.starts_with(package) && mapped_frame.in_app.is_none() {
-                    mapped_frame.in_app = Some(true);
+        // Fix up the frames' in-app fields only if they were actually mapped
+        if let Some(frames) = frames.as_mut() {
+            for frame in frames {
+                // mark the frame as in_app after deobfuscation based on the release package name
+                // only if it's not present
+                if let Some(package) = release_package {
+                    if frame.module.starts_with(package) && frame.in_app.is_none() {
+                        frame.in_app = Some(true);
+                    }
                 }
             }
-
-            return vec![mapped_frame];
         }
 
-        // Return the raw frame if remapping didn't work
-        vec![frame.clone()]
+        // If all else fails, just return the original frame.
+        let mut frames = frames.unwrap_or_else(|| vec![frame.clone()]);
+
+        for frame in &mut frames {
+            // add the signature if we received one and we were
+            // able to translate/deobfuscate it
+            if let Some(signature) = &deobfuscated_signature {
+                frame.signature = Some(signature.format_signature());
+            }
+        }
+
+        frames
+    }
+
+    /// Tries to remap a `JvmFrame` using a `proguard::StackFrame`
+    /// constructed from it.
+    ///
+    /// The `buf` parameter is used as a buffer for the frames returned
+    /// by `remap_frame`.
+    fn map_full_frame<'a>(
+        mapper: &'a proguard::ProguardMapper<'a>,
+        original_frame: &JvmFrame,
+        proguard_frame: &proguard::StackFrame<'a>,
+        buf: &mut Vec<proguard::StackFrame<'a>>,
+    ) -> Option<Vec<JvmFrame>> {
+        buf.clear();
+        buf.extend(mapper.remap_frame(proguard_frame));
+
+        if buf.is_empty() {
+            return None;
+        }
+
+        let bottom_class = buf[buf.len() - 1].class();
+
+        // sentry expects stack traces in reverse order
+        let res = buf
+            .iter()
+            .rev()
+            .map(|new_frame| {
+                let mut mapped_frame = JvmFrame {
+                    module: new_frame.class().to_owned(),
+                    function: new_frame.method().to_owned(),
+                    lineno: Some(new_frame.line() as u32),
+                    ..original_frame.clone()
+                };
+
+                // clear the filename for all *foreign* classes
+                if mapped_frame.module != bottom_class {
+                    mapped_frame.filename = None;
+                    mapped_frame.abs_path = None;
+                }
+
+                mapped_frame
+            })
+            .collect();
+        Some(res)
+    }
+
+    /// Tries to remap a frame's class and method.
+    fn map_class_method(
+        mapper: &proguard::ProguardMapper,
+        frame: &JvmFrame,
+    ) -> Option<Vec<JvmFrame>> {
+        let (mapped_class, mapped_method) = mapper.remap_method(&frame.module, &frame.function)?;
+
+        Some(vec![JvmFrame {
+            module: mapped_class.to_owned(),
+            function: mapped_method.to_owned(),
+            ..frame.clone()
+        }])
+    }
+
+    /// Tries to remap a frame's class.
+    fn map_class(mapper: &proguard::ProguardMapper, frame: &JvmFrame) -> Option<Vec<JvmFrame>> {
+        let mapped_class = mapper.remap_class(&frame.module)?;
+
+        Some(vec![JvmFrame {
+            module: mapped_class.to_owned(),
+            ..frame.clone()
+        }])
     }
 
     /// Applies source context from the given list of source bundles to a frame.
@@ -278,9 +355,10 @@ impl ProguardService {
     /// If one of the source bundles contains the correct file name, we apply it, otherwise
     /// the frame stays unmodified.
     fn apply_source_context(source_bundles: &[SourceBundleDebugSession<'_>], frame: &mut JvmFrame) {
-        let Some(lineno) = frame.lineno else {
+        let lineno = match frame.lineno {
             // can't apply source context without line number
-            return;
+            None | Some(0) => return,
+            Some(n) => n,
         };
 
         let source_file_name = build_source_file_name(frame);
@@ -294,7 +372,7 @@ impl ProguardService {
             };
 
             if let Some((pre_context, context_line, post_context)) =
-                get_context_lines(contents, lineno as usize, None, None)
+                get_context_lines(contents, lineno as usize, 0, None)
             {
                 frame.pre_context = pre_context;
                 frame.context_line = Some(context_line);
@@ -332,7 +410,7 @@ fn build_source_file_name(frame: &JvmFrame) -> String {
             source_file_name.push_str(abs_path_before_dot);
         }
         None => {
-            let module_before_dollar = module.rsplit_once('$').map(|p| p.0).unwrap_or(module);
+            let module_before_dollar = module.split_once('$').map(|p| p.0).unwrap_or(module);
             source_file_name.push_str(&module_before_dollar.replace('.', "/"));
         }
     };
@@ -414,17 +492,45 @@ io.sentry.sample.MainActivity -> io.sentry.sample.MainActivity:
                 index: 1,
                 ..Default::default()
             },
+            JvmFrame {
+                function: "t".to_owned(),
+                module: "io.sentry.sample.MainActivity".to_owned(),
+                signature: Some("(Landroid/view/View;)V".to_owned()),
+                index: 2,
+                ..Default::default()
+            },
+            JvmFrame {
+                // this function map is onClickHandler(android.view.View):40 -> t
+                // not onClickHandler(android.view.View):40 -> onClickHandler
+                // hence the class remapping should fail,
+                // but the signature should still be properly translated to java type
+                function: "onClickHandler".to_owned(),
+                module: "io.sentry.sample.MainActivity".to_owned(),
+                signature: Some("(Landroid/view/View;)V".to_owned()),
+                index: 3,
+                ..Default::default()
+            },
+            JvmFrame {
+                // this module (Class) does not exist in the mapping,
+                // hence the whole frame remapping should fail,
+                // but the signature should still be properly translated to java type
+                function: "onClickHandler".to_owned(),
+                module: "io.sentry.sample.ClassDoesNotExist".to_owned(),
+                signature: Some("(Landroid/view/View;)V".to_owned()),
+                index: 4,
+                ..Default::default()
+            },
         ];
 
         let mapping = ProguardMapping::new(proguard_source);
-        let mapper = ProguardMapper::new(mapping);
+        let mapper = ProguardMapper::new_with_param_mapping(mapping, true);
 
         let mapped_frames: Vec<_> = frames
             .iter()
             .flat_map(|frame| ProguardService::map_frame(&[&mapper], frame, None).into_iter())
             .collect();
 
-        assert_eq!(mapped_frames.len(), 4);
+        assert_eq!(mapped_frames.len(), 7);
 
         assert_eq!(mapped_frames[0].function, "onClick");
         assert_eq!(
@@ -454,6 +560,24 @@ io.sentry.sample.MainActivity -> io.sentry.sample.MainActivity:
         );
         assert_eq!(mapped_frames[3].module, "io.sentry.sample.MainActivity");
         assert_eq!(mapped_frames[3].index, 1);
+
+        assert_eq!(mapped_frames[4].function, "onClickHandler");
+        assert_eq!(
+            mapped_frames[4].signature.as_ref().unwrap(),
+            "(android.view.View)"
+        );
+
+        assert_eq!(mapped_frames[5].function, "onClickHandler");
+        assert_eq!(
+            mapped_frames[5].signature.as_ref().unwrap(),
+            "(android.view.View)"
+        );
+
+        assert_eq!(mapped_frames[6].function, "onClickHandler");
+        assert_eq!(
+            mapped_frames[6].signature.as_ref().unwrap(),
+            "(android.view.View)"
+        );
     }
 
     // based on the Python test `test_sets_inapp_after_resolving`.
@@ -523,5 +647,151 @@ org.slf4j.helpers.Util$ClassContext -> org.a.b.g$b:
         // this must happen somewhere else in `sentry`.
         // assert_eq!(mapped_frames[3].in_app, Some(false));
         assert_eq!(mapped_frames[4].in_app, Some(true));
+    }
+
+    #[test]
+    fn doesnt_set_in_app_if_not_resolved() {
+        let frame = JvmFrame {
+            function: "main".into(),
+            module: "android.app.ActivityThread".into(),
+            lineno: Some(8918),
+            ..Default::default()
+        };
+
+        let remapped = ProguardService::map_frame(&[], &frame, Some("android"));
+
+        assert_eq!(remapped.len(), 1);
+        // The frame didn't get mapped, so we shouldn't set `in_app` even though
+        // the condition is satisfied.
+        assert!(remapped[0].in_app.is_none());
+    }
+
+    #[test]
+    fn line_0_1() {
+        let proguard_source = br#"com.example.App -> com.example.App:
+# {"id":"sourceFile","fileName":"App.java"}
+    boolean injected -> g
+    foo.bar.android.internal.managers.ApplicationComponentManager componentManager -> h
+    0:3:void <init>():18:18 -> <init>
+    4:5:void <init>():19:19 -> <init>
+    6:18:void <init>():21:21 -> <init>
+    1:1:foo.bar.internal.GeneratedComponentManager componentManager():17:17 -> componentManager
+    2:2:foo.bar.android.internal.managers.ApplicationComponentManager componentManager():31:31 -> componentManager
+    0:6:java.lang.Object generatedComponent():36:36 -> generatedComponent
+    0:4:void barInternalInject():47:47 -> onCreate
+    0:4:void onCreate():42 -> onCreate
+    5:6:void barInternalInject():48:48 -> onCreate
+    5:6:void onCreate():42 -> onCreate
+    7:12:java.lang.Object generatedComponent():36:36 -> onCreate
+    7:12:void barInternalInject():51 -> onCreate
+    7:12:void onCreate():42 -> onCreate
+    13:20:void barInternalInject():51:51 -> onCreate
+    13:20:void onCreate():42 -> onCreate
+    21:24:void onCreate():43:43 -> onCreate
+"#;
+
+        let mapping = ProguardMapping::new(proguard_source);
+        let mapper = ProguardMapper::new(mapping);
+
+        let frame = JvmFrame {
+            function: "onCreate".into(),
+            module: "com.example.App".into(),
+            index: 0,
+            ..Default::default()
+        };
+
+        let mapped_frames = ProguardService::map_frame(&[&mapper], &frame, None);
+
+        assert_eq!(mapped_frames.len(), 2);
+
+        assert_eq!(
+            mapped_frames[0],
+            JvmFrame {
+                function: "onCreate".into(),
+                module: "com.example.App".into(),
+                lineno: Some(0),
+                index: 0,
+                ..Default::default()
+            }
+        );
+
+        // Without the "line 0" change, this frame doesn't exist.
+        // The `retrace` implementation at
+        // https://dl.google.com/android/repository/commandlinetools-mac-11076708_latest.zip
+        // also returns this, no matter whether you give it line 0 or no line at all.
+        assert_eq!(
+            mapped_frames[1],
+            JvmFrame {
+                function: "barInternalInject".into(),
+                module: "com.example.App".into(),
+                lineno: Some(0),
+                index: 0,
+                ..Default::default()
+            }
+        );
+    }
+
+    #[test]
+    fn line_0_2() {
+        let proguard_source = br#"com.google.firebase.concurrent.CustomThreadFactory$$ExternalSyntheticLambda0 -> com.google.firebase.concurrent.a:
+# {"id":"sourceFile","fileName":"R8$$SyntheticClass"}
+# {"id":"com.android.tools.r8.synthesized"}
+    int com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.$r8$classId -> d
+    com.google.firebase.concurrent.CustomThreadFactory com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.f$0 -> e
+    java.lang.Runnable com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.f$1 -> f
+    0:9:void com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.<init>(com.google.firebase.concurrent.CustomThreadFactory,java.lang.Runnable):0:0 -> <init>
+    0:9:void com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.$r8$init$synthetic(java.lang.Object,java.lang.Object,int):0 -> <init>
+      # {"id":"com.android.tools.r8.synthesized"}
+      # {"id":"com.android.tools.r8.residualsignature","signature":"(ILjava/lang/Object;Ljava/lang/Object;)V"}
+    0:25:void com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0.run$bridge():0:0 -> run
+      # {"id":"com.android.tools.r8.synthesized"}
+y.b -> y.b:
+# {"id":"sourceFile","fileName":"FutureExt.kt"}
+    0:4:void a(com.google.common.util.concurrent.ListenableFuture,com.drivit.core.DrivitCloud$OperationListener):1:1 -> a
+    5:8:void a(com.google.common.util.concurrent.ListenableFuture,com.drivit.core.DrivitCloud$OperationListener):2:2 -> a
+"#;
+
+        let mapping = ProguardMapping::new(proguard_source);
+        let mapper = ProguardMapper::new(mapping);
+
+        let frame = JvmFrame {
+            function: "run".into(),
+            module: "com.google.firebase.concurrent.a".into(),
+            index: 0,
+            ..Default::default()
+        };
+
+        let mapped_frames = ProguardService::map_frame(&[&mapper], &frame, None);
+
+        assert_eq!(mapped_frames.len(), 1);
+
+        assert_eq!(
+            mapped_frames[0],
+            JvmFrame {
+                function: "run$bridge".into(),
+                // Without the "line 0" change, this is "com.google.firebase.concurrent.CustomThreadFactory$$ExternalSyntheticLambda0".
+                // The `retrace` implementation at
+                // https://dl.google.com/android/repository/commandlinetools-mac-11076708_latest.zip
+                // also returns this, no matter whether you give it line 0 or no line at all.
+                module: "com.google.firebase.concurrent.CustomThreadFactory$$InternalSyntheticLambda$1$53203795c28a6fcdb3bac755806c9ee73cb3e8dcd4c9bbf8ca5d25d4d9c378dd$0".into(),
+                lineno: Some(0),
+                index: 0,
+                ..Default::default()
+            }
+        );
+    }
+
+    #[test]
+    fn test_build_source_file_name() {
+        let frame = JvmFrame {
+            function: "run".into(),
+            module: "com.foo.bar.ui.activities.base.BaseActivity$$ExternalSyntheticLambda2".into(),
+            ..Default::default()
+        };
+
+        assert_eq!(
+            build_source_file_name(&frame),
+            "~/com/foo/bar/ui/activities/base/BaseActivity.jvm"
+        );
     }
 }